deleted actions folder, replaced main actions to sagas.js
