--- src/models/themes/actions/index.js
@@ -1,5 +1,0 @@
-export { getThemes, loadingThemesStarted, loadingThemesSucceed, loadingThemesFailed } from './actions-get.js';
-export { createTheme, creatingThemeStarted, creatingThemeSucceed, creatingThemeFailed } from './actions-post.js';
-export { editTheme, editingThemeStarted, editingThemeSucceed, editingThemeFailed } from './actions-put.js';
-export { deleteTheme, deletingThemeStarted, deletingThemeSucceed, deletingThemeFailed } from './actions-delete.js'
-
