+++ src/models/themes/index.js
@@ -1,4 +1,3 @@
 export { themesReducer } from './reducer.js';
-export { themesWatcher } from './sagas.js';
+export { themesWatcher, getThemes, createTheme, editTheme, deleteTheme } from './sagas.js';
 export { themesDataSelector, themesIsLoadingSelector } from './selectors.js';
-export { getThemes, createTheme, editTheme, deleteTheme } from './actions/index.js';
