+++ src/models/themes/types.js
@@ -1,4 +1,4 @@
-export const GET_THEMES = 'THEMES/GET_THEMES';
+export const FETCH_THEMES = 'THEMES/FETCH_THEMES';
 export const LOADING_THEMES_STARTED = 'THEMES/LOADING_STARTED'
 export const LOADING_THEMES_SUCCESS = 'THEMES/LOADING_FINISHED';
 export const LOADING_THEMES_FAILED = 'THEMES/LOADING_FAILED';
